Revert -Werr. Fix a lot of warnings.
Mainly related to signed / unsigned comparisions.

+++ src/com/MPIDirectCommunication.cpp
@@ -31,8 +31,7 @@ MPIDirectCommunication::~MPIDirectCommunication() {
   closeConnection();
 }
 
-int
-MPIDirectCommunication::getRemoteCommunicatorSize() {
+size_t MPIDirectCommunication::getRemoteCommunicatorSize() {
   preciceTrace("getRemoteCommunicatorSize()");
   assertion(isConnected());
   int remoteSize = 0;
